INNOW-3: some minor changes that implement logic for kafka producer

+++ src/main/java/by/example/bookstore_api/model/dto/request/OrderRequestDto.java
@@ -5,6 +5,6 @@ import java.util.UUID;
 public record OrderRequestDto(
         UUID userId,
         UUID bookId,
-        long quantity
+        int quantity
 ) {
 }
